Initial Dashboard date range dropdown implementation
